refactor(proposal): Response changed in CreateQuoteProposalcontroller, and changes in Proposal namespace.
