+++ app/Http/Controllers/Quotes/Proposals/GetAllQuoteProposalsController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Quotes\Proposal;
+namespace App\Http\Controllers\Quotes\Proposals;
 
 use App\Constants\UserTypesAbilitiesConstant;
 use App\Http\Controllers\Controller;
