--- routes/api/quote_proposals.php
@@ -1,16 +1,0 @@
-<?php
-
-
-use App\Http\Controllers\QuoteProposal\CreateQuoteProposalController;
-use App\Http\Controllers\Quotes\Proposal\GetAllQuoteProposalsController;
-use App\Http\Controllers\Quotes\Proposal\ReviewQuoteProposalController;
-
-
-Route::post('/', CreateQuoteProposalController::class)
-    ->middleware(['auth:sanctum', 'abilities:is-user']);
-
-
-Route::middleware(['auth:sanctum', 'abilities:is-admin'])->group(function () {
-    Route::get('/', GetAllQuoteProposalsController::class);
-    Route::patch('/{quote_proposal}', ReviewQuoteProposalController::class);
-});
